🚚 Move FileList component to /documents/components
